rgbh: add a RGBManipulator class
This new class is initialized with an RGB value which can then be
transformed by incrementing/decrementing value/chroma and rotating the
hue angle.  In each case, the other characteristics are maintained
except for hue rotation where chroma is maintained but value may vary
due the "the way the world works". Each transformation method returns
a boolean value to indicate whether a change was actually made or was
refused because it was not possible.

This is intended for use in tweaking RGBs in the sample matcher.

Signed-off-by: Peter Williams <pwil3058@gmail.com>